Refactor Movie Screen into separate package
Also, split movie model into separate class
