+++ composeApp/src/androidMain/kotlin/io/github/couchtracker/App.kt
@@ -20,7 +20,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import io.github.couchtracker.ui.AnimationDefaults
 import io.github.couchtracker.ui.screens.main.MainScreen
-import io.github.couchtracker.ui.screens.movieScreen
+import io.github.couchtracker.ui.screens.movie.movieScreen
 import org.koin.compose.KoinContext
 import kotlin.math.roundToInt
