+++ composeApp/src/androidMain/kotlin/io/github/couchtracker/ui/screens/movie/MovieScreen.kt
@@ -1,8 +1,7 @@
 @file:OptIn(ExperimentalMaterial3Api::class, ExperimentalLayoutApi::class)
 
-package io.github.couchtracker.ui.screens
+package io.github.couchtracker.ui.screens.movie
 
-import android.content.Context
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
@@ -24,7 +23,6 @@ import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ColorScheme
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
@@ -57,21 +55,14 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import androidx.core.graphics.drawable.toBitmap
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import androidx.palette.graphics.Palette
 import app.moviebase.tmdb.image.TmdbImageType
 import app.moviebase.tmdb.image.TmdbImageUrlBuilder
-import app.moviebase.tmdb.model.TmdbCrew
-import app.moviebase.tmdb.model.TmdbFileImage
-import app.moviebase.tmdb.model.TmdbGenre
 import coil.compose.AsyncImage
-import coil.imageLoader
-import coil.request.ImageRequest
 import couch_tracker_app.composeapp.generated.resources.Res
 import couch_tracker_app.composeapp.generated.resources.back_action
 import couch_tracker_app.composeapp.generated.resources.movie_by_director
@@ -81,28 +72,16 @@ import io.github.couchtracker.db.tmdbCache.TmdbCache
 import io.github.couchtracker.db.user.movie.ExternalMovieId
 import io.github.couchtracker.db.user.movie.TmdbExternalMovieId
 import io.github.couchtracker.db.user.movie.UnknownExternalMovieId
-import io.github.couchtracker.tmdb.TmdbException
 import io.github.couchtracker.tmdb.TmdbLanguage
 import io.github.couchtracker.tmdb.TmdbMovie
-import io.github.couchtracker.tmdb.TmdbRating
-import io.github.couchtracker.tmdb.rating
 import io.github.couchtracker.ui.components.BackgroundTopAppBar
 import io.github.couchtracker.ui.components.DefaultErrorScreen
 import io.github.couchtracker.ui.components.LoadableScreen
-import io.github.couchtracker.ui.generateColorScheme
-import io.github.couchtracker.ui.screens.main.MOVIE_COLOR_SCHEME
 import io.github.couchtracker.utils.Loadable
 import io.github.couchtracker.utils.formatAndList
 import io.github.couchtracker.utils.str
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
-import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import org.koin.compose.koinInject
-import kotlin.coroutines.CoroutineContext
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.minutes
 
 private const val ARGUMENTS_MOVIE_ID = "movieId"
 private const val ARGUMENTS_LANGUAGE = "language"
@@ -137,20 +116,6 @@ fun NavGraphBuilder.movieScreen() {
         }
     }
 }
-
-private data class MovieScreenModel(
-    val title: String,
-    val tagline: String,
-    val overview: String,
-    val year: Int?,
-    val runtime: Duration?,
-    val rating: TmdbRating?,
-    val genres: List<TmdbGenre>,
-    val director: List<TmdbCrew>,
-    val images: List<TmdbFileImage>,
-    val backdrop: ImageRequest?,
-    val colorScheme: ColorScheme,
-)
 
 @Composable
 fun MovieScreen(movie: TmdbMovie) {
@@ -359,66 +324,6 @@ private fun LazyListScope.tagsComposable(tags: List<String>) {
                     Text(tag, style = MaterialTheme.typography.labelLarge)
                 }
             }
-        }
-    }
-}
-
-private suspend fun loadMovie(
-    ctx: Context,
-    tmdbCache: TmdbCache,
-    movie: TmdbMovie,
-    width: Int,
-    height: Int,
-    coroutineContext: CoroutineContext = Dispatchers.Default,
-): Loadable<MovieScreenModel> = coroutineScope {
-    withContext(coroutineContext) {
-        try {
-            val credits = async { movie.credits(tmdbCache) }
-            val images = async { movie.images(tmdbCache) }
-            val details = movie.details(tmdbCache)
-            val backdropImage = details.backdropImage
-            val backdropImageRequest: ImageRequest?
-            var colorScheme: ColorScheme = MOVIE_COLOR_SCHEME
-            if (backdropImage != null) {
-                val url = TmdbImageUrlBuilder.build(backdropImage, width, height)
-                backdropImageRequest = ImageRequest.Builder(ctx)
-                    // Necessary for palette generation
-                    .allowHardware(false)
-                    .data(url)
-                    .size(width, height)
-                    .build()
-                val drawable = ctx.imageLoader.execute(backdropImageRequest).drawable
-                if (drawable != null) {
-                    colorScheme = async {
-                        val bitmap = drawable.toBitmap()
-                        bitmap.prepareToDraw()
-                        val palette = Palette.Builder(bitmap).generate()
-                        palette.generateColorScheme()
-                    }.await()
-                }
-            } else {
-                backdropImageRequest = null
-            }
-            Loadable.Loaded(
-                MovieScreenModel(
-                    title = details.title,
-                    tagline = details.tagline,
-                    overview = details.overview,
-                    year = details.releaseDate?.year,
-                    runtime = details.runtime?.minutes,
-                    rating = details.rating(),
-                    genres = details.genres,
-                    director = credits.await().crew.filter { it.job == "Director" },
-                    backdrop = backdropImageRequest,
-                    images = images.await().let { imgs ->
-                        (imgs.backdrops + imgs.posters).sortedByDescending { it.voteAverage }
-                    },
-                    colorScheme = colorScheme,
-                ),
-            )
-        } catch (e: TmdbException) {
-            // TODO: translate
-            Loadable.Error(e.message ?: "Error")
         }
     }
 }
